requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0

# HTTP Requests
requests>=2.31.0

# Vector Database
chromadb>=0.4.0

# LLM Providers
openai>=1.0.0
anthropic>=0.18.0

# Configuration & Validation
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Progress Bars
tqdm>=4.66.0

# Markdown Rendering
markdown>=3.5.0

# Optional: Faster XML parsing (3-5x speedup)
# Uncomment to enable:
lxml>=5.0.0
markdown>=3.5.0
